+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- **Jumping input widthn** - Issue #1, number controls jumped as text width changed
+
 ## [0.6.0] - 2026-01-12
 
 ### Added