+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.7.0] - 2026-02-11
 
 ### Added
 - **Separator** - Added separator control to visually separate groups of related controls
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Set to `block` for full-width layout
 
 ### Fixed
-- **Jumping input widthn** - Issue #1, number controls jumped as text width
+- **Jumping input width** - Issue #1, number controls jumped as text width
   changed
 - **Setting array items** - Setting array items, like `colors[0]` caused the
   existing other items to get lost